+++ e2e/cypress/support/commands.js
@@ -14,6 +14,7 @@ import {
 import { defaultConfig } from '../testAppConfig';
 import { fireEvent } from '@testing-library/react';
 import RicosDriver from '../../../packages/ricos-driver/web/src/RicosDriver';
+import { merge } from 'lodash';
 // Viewport size commands
 const resizeForDesktop = () => cy.viewport('macbook-15');
 const resizeForMobile = () => cy.viewport('iphone-6');
@@ -91,8 +92,8 @@ Cypress.Commands.add('loadEditorAndViewer', (fixtureName, config) =>
 Cypress.Commands.add('loadIsolatedEditorAndViewer', fixtureName =>
   run('rce-isolated', fixtureName)
 );
-Cypress.Commands.add('loadRicosEditorAndViewer', (fixtureName, config) =>
-  run('ricos', fixtureName, config)
+Cypress.Commands.add('loadRicosEditorAndViewer', (fixtureName, ...config) =>
+  run('ricos', fixtureName, merge(...config))
 );
 
 Cypress.Commands.add('loadTestAppOnSsr', (fixtureName, compName) => {