+++ e2e/test-env/src/client/RicosTestApp/RicosTestApp.jsx
@@ -4,6 +4,7 @@ import windowContentStateHoc from '../WindowContentStateHoc';
 import { RichContentEditor } from 'wix-rich-content-editor';
 import { RicosEditor } from 'ricos-editor';
 import { RicosViewer } from 'ricos-viewer';
+import { createTheme } from 'ricos-theme';
 import { default as editorPlugins } from './editorPlugins';
 import { default as viewerPlugins } from './viewerPlugins';
 import './styles.global.scss';
@@ -11,11 +12,16 @@ import theme from '../../../../../examples/main/shared/theme/theme';
 import { testVideos } from '../../../../../examples/main/shared/utils/mock';
 import { createPreview } from 'wix-rich-content-preview';
 import { TextSelectionToolbar, TwitterButton } from 'wix-rich-content-text-selection-toolbar';
-import { FORMATTING_BUTTONS, TOOLBARS } from 'wix-rich-content-editor-common';
+import { TOOLBARS } from 'wix-rich-content-editor-common';
+import { wixPalettes } from '../../../../tests/resources/palettesExample';
 
 const onVideoSelected = (url, updateEntity) => {
   setTimeout(() => updateEntity(testVideos[1]), 1);
 };
+const determinePalette = paletteType =>
+  paletteType ? (paletteType === 'light' ? wixPalettes[1] : wixPalettes[9]) : undefined;
+const setBackground = palette => (palette ? { backgroundColor: palette[5].value } : {});
+const setForeground = palette => (palette ? { color: palette[9].value } : {});
 class RicosTestApp extends PureComponent {
   constructor(props) {
     super(props);
@@ -24,7 +30,7 @@ class RicosTestApp extends PureComponent {
 
   renderEditor = () => {
     const createToolbarSettings = (addPluginMenuConfig, footerToolbarConfig) => ({
-      getToolbarSettings: ({ textButtons }) => [
+      getToolbarSettings: () => [
         { name: TOOLBARS.SIDE, addPluginMenuConfig },
         {
           name: TOOLBARS.MOBILE,
@@ -36,6 +42,8 @@ class RicosTestApp extends PureComponent {
 
     const { contentState, onEditorChange, locale, isMobile, testAppConfig = {} } = this.props;
     const { addPluginMenuConfig, footerToolbarConfig } = testAppConfig.toolbarConfig || {};
+    const { skipCssOverride, paletteType } = testAppConfig.theme || {};
+    const palette = determinePalette(paletteType);
     return (
       <RicosEditor
         plugins={editorPlugins(testAppConfig.plugins)}
@@ -43,7 +51,8 @@ class RicosTestApp extends PureComponent {
         content={contentState}
         isMobile={isMobile}
         locale={locale}
-        cssOverride={theme}
+        theme={palette && createTheme({ palette })}
+        cssOverride={!skipCssOverride && theme}
         toolbarSettings={createToolbarSettings(addPluginMenuConfig, footerToolbarConfig)}
       >
         <RichContentEditor
@@ -57,15 +66,17 @@ class RicosTestApp extends PureComponent {
   };
 
   renderViewer = () => {
-    const { isMobile, contentState, locale, seoMode, testAppConfig } = this.props;
-
+    const { isMobile, contentState, locale, seoMode, testAppConfig = {} } = this.props;
+    const { skipCssOverride, paletteType } = testAppConfig.theme || {};
+    const palette = determinePalette(paletteType);
     return (
       <RicosViewer
         plugins={viewerPlugins(testAppConfig.plugins)}
         content={contentState}
         isMobile={isMobile}
         locale={locale}
-        cssOverride={theme}
+        theme={palette && createTheme({ palette })}
+        cssOverride={!skipCssOverride && theme}
         seoSettings={seoMode}
         preview={testAppConfig.showDefaultPreview && createPreview()}
       />
@@ -73,17 +84,19 @@ class RicosTestApp extends PureComponent {
   };
 
   render() {
-    const { isMobile } = this.props;
+    const { isMobile, testAppConfig = {} } = this.props;
+    const { theme: { paletteType } = {} } = testAppConfig;
+    const palette = determinePalette(paletteType);
     return (
-      <div className={`testApp ${isMobile ? 'mobile' : ''}`}>
+      <div className={`testApp ${isMobile ? 'mobile' : ''}`} style={setBackground(palette)}>
         <div>
-          <h3>Editor</h3>
+          <h3 style={setForeground(palette)}>Editor</h3>
           <div className="rcWrapper rce" id="RicosEditorContainer" data-hook="ricos-editor">
             {this.renderEditor()}
           </div>
         </div>
         <div>
-          <h3>Viewer</h3>
+          <h3 style={setForeground(palette)}>Viewer</h3>
           <div
             className="rcWrapper rcv"
             id="RicosViewerContainer"