+++ examples/storybook/stories/TestCase/VeryBigPost.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { RichContentEditorBox, Page } from '../Components/StoryParts';
 import EditorWrapper from '../Components/EditorWrapper';
 import fixture from '../../../../e2e/tests/fixtures/very-big-post.json';
-import { wixPalettes } from '../palettesExample';
+import { wixPalettes } from '../../../../e2e/tests/resources/palettesExample';
 
 export default () => {
   return (